+++ spec/rubocop/cop/style/block_comments_spec.rb
@@ -18,4 +18,28 @@ describe Rubocop::Cop::Style::BlockComments do
                    ['# comment'])
     expect(cop.offenses).to be_empty
   end
+
+  it 'auto-corrects a block comment into a regular comment' do
+    new_source = autocorrect_source(cop, ['=begin',
+                                          'comment line 1',
+                                          '',
+                                          'comment line 2',
+                                          '=end',
+                                          'def foo',
+                                          'end'])
+    expect(new_source).to eq(['# comment line 1',
+                              '#',
+                              '# comment line 2',
+                              'def foo',
+                              'end'].join("\n"))
+  end
+
+  it 'auto-corrects an empty block comment by removing it' do
+    new_source = autocorrect_source(cop, ['=begin',
+                                          '=end',
+                                          'def foo',
+                                          'end'])
+    expect(new_source).to eq(['def foo',
+                              'end'].join("\n"))
+  end
 end